# specs/002-phase-ii-web-app/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo Management API
  description: REST API for the Todo Management System
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.todoapp.com
    description: Production server

paths:
  /auth/register:
    post:
      summary: Register a new user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/login:
    post:
      summary: Login user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/logout:
    post:
      summary: Logout user
      tags:
        - Authentication
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
        '401':
          description: Unauthorized

  /todos:
    get:
      summary: Get all todos for the authenticated user
      tags:
        - Todos
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of todos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TodoResponse'
        '401':
          description: Unauthorized
    post:
      summary: Create a new todo
      tags:
        - Todos
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoRequest'
      responses:
        '201':
          description: Todo created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized

  /todos/{id}:
    get:
      summary: Get a specific todo
      tags:
        - Todos
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Todo details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '401':
          description: Unauthorized
        '404':
          description: Todo not found
    put:
      summary: Update a specific todo
      tags:
        - Todos
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTodoRequest'
      responses:
        '200':
          description: Todo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
        '404':
          description: Todo not found
    delete:
      summary: Delete a specific todo
      tags:
        - Todos
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Todo deleted successfully
        '401':
          description: Unauthorized
        '404':
          description: Todo not found

  /todos/{id}/toggle-complete:
    patch:
      summary: Toggle todo completion status
      tags:
        - Todos
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Todo completion status toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '401':
          description: Unauthorized
        '404':
          description: Todo not found

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    UserRegistration:
      type: object
      required:
        - email
        - password
        - first_name
        - last_name
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          minLength: 8
          example: "securePassword123"
        first_name:
          type: string
          maxLength: 50
          example: "John"
        last_name:
          type: string
          maxLength: 50
          example: "Doe"

    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          example: "securePassword123"

    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          example: "user@example.com"
        first_name:
          type: string
          example: "John"
        last_name:
          type: string
          example: "Doe"
        created_at:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"

    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          example: "bearer"
        user:
          $ref: '#/components/schemas/UserResponse'

    CreateTodoRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          maxLength: 200
          example: "Buy groceries"
        description:
          type: string
          maxLength: 1000
          example: "Milk, eggs, bread, fruits"
        is_completed:
          type: boolean
          default: false

    UpdateTodoRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
          example: "Buy groceries"
        description:
          type: string
          maxLength: 1000
          example: "Milk, eggs, bread, fruits"
        is_completed:
          type: boolean
          example: false

    TodoResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        title:
          type: string
          example: "Buy groceries"
        description:
          type: string
          example: "Milk, eggs, bread, fruits"
        is_completed:
          type: boolean
          example: false
        user_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440001"
        created_at:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"

    Error:
      type: object
      properties:
        detail:
          type: string
          example: "An error occurred"